+++ CHANGELOG.md
@@ -4,7 +4,12 @@ NOWnewsAdmin 的所有改變將在此 CHANGELOG 文件中記錄。
 格式基於 [Keep a Changelog](http://keepachangelog.com/zh-TW/0.3.0/)
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
-## 1.0.2 - 2017-07-05
+## 1.0.4 - 2017-07-06
+### Added
+-  CKeditor加上符號選擇器 @appleoxxo
+-  CKeditor開啟預設ACF過濾 @appleoxxo
+
+## 1.0.3 - 2017-07-05
 ### Bug
 -  修正編輯器無法複製圖片問題 @wb @appleoxxo
 
@@ -13,11 +18,11 @@ NOWnewsAdmin 的所有改變將在此 CHANGELOG 文件中記錄。
 -  新聞內容若有MLB的ifrme http連結改成https @appleoxxo
 
 ## 1.0.1 - 2017-07-03
-### Change
+### Changed
 -  新聞列表加上PV欄位 @appleoxxo
 
 ## 1.0.0 - 2017-06-28
-### Change
+### Changed
 -  正式上線 @appleoxxo
 
 ## 0.0.1 - 2016-11-08